% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bencapex}
\alias{bencapex}
\title{A Worked Example from Benati and Capurri (2026)}
\format{
A data frame with 240 observations on the following 2 variables.
\describe{
\item{\code{rowv}}{how the row voter voted on a resolution}
\item{\code{colv}}{how the column voter voted on a resolution}
}
}
\usage{
bencapex
}
\description{
A simple worked example illustrating what the authors call an "alignment
index." You can think of this as a kind of chance-corrected \emph{S} score.
}
\details{
This is an expansion of Table 1 in their article. Valid vote values
identified are 1 (yes), 2 (abstain), and 3 (no).
}
\references{
Benati, Stefano, and Agnese Capurri. 2026. "The Alignment index and its
application to voting at the United Nations General Assembly." \emph{Quality &
Quantity}. \doi{10.1007/s11135-026-02814-x}
}
\keyword{datasets}
