% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmyrus14}
\alias{gmyrus14}
\title{Select Alliance Portfolios of Germany and Russia, 1914}
\format{
A data frame with 20 observations on the following 4 variables.
\describe{
\item{\code{state}}{a three-character code indicating a state}
\item{\code{syscap}}{the capabilities of the state, as a potential weight}
\item{\code{gmy}}{the alliance commitment for Germany with the state identified in the \code{state} column}
\item{\code{rus}}{the alliance commitment for Russia with the state identified in the \code{state} column}
}
}
\usage{
gmyrus14
}
\description{
A simple example of alliance portfolios of Germany and Russia, by way of
Signorino and Ritter's (1999) example.
}
\details{
The data come by way of Signorino and Ritter (1999, Table 6).
}
\references{
Signorino, Curtis S. and Jeffrey M. Ritter. "Tau-b or Not Tau-B: Measuring
the Similarity of Foreign Policy Positions." \emph{International Studies Quarterly}
43(1): 115–44.
}
\keyword{datasets}
