% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usamex46}
\alias{usamex46}
\title{American-Mexican Dyadic Voting Patterns in the United Nations, 1946}
\format{
A data frame with 38 observations on the following 6 variables.
\describe{
\item{\code{resid}}{an identifier for a roll-call vote ID}
\item{\code{ccode1}}{the Correlates of War state code for the United States (2)}
\item{\code{ccode2}}{the Correlates of War state code for Mexico (70)}
\item{\code{year}}{a numeric constant for the year (1946)}
\item{\code{vote1}}{an integer for how the United States voted on the resolution identified in the \code{resid} column}
\item{\code{vote2}}{an integer for how Mexico voted on the resolution identified in the \code{resid} column}
}
}
\usage{
usamex46
}
\description{
A simple example of voting patterns for the United States and Mexico in the
United Nations in 1946.
}
\details{
Data are from a June 2024 of the United Nations voting data provided by
Erik Voeten on his Dataverse for the project.

Valid vote values identified are 1 (yes), 2 (abstain), and 3 (no).
}
\references{
Bailey, Michael A., Anton Strezhnev, and Erik Voeten. 2017. "Estimating
Dynamic State Preferences from United Nations Voting Data." \emph{Journal of
Conflict Resolution} 61(2): 430-56.
}
\keyword{datasets}
