% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcai.R
\name{bcai}
\alias{bcai}
\title{Calculate Benati and Capurri's (2026) alignment index (A)}
\usage{
bcai(x1, x2, distances = "absolute", weights = NULL, levels = NULL)
}
\arguments{
\item{x1}{a vector, and one assumes an integer}

\item{x2}{a vector, and one assumes an integer}

\item{distances}{the type of distances between ratings/attachments to
estimate. Can be either "absolute" or "squared". Defaults to "absolute", but
see note in details section.}

\item{weights}{a vector of weights. Defaults to NULL for creating unweighted
A index values.}

\item{levels}{defaults to NULL, but an optional vector that defines the full
sequence of values that could be observed in \code{x1} and \code{x2}. If NULL, the
function looks for observed values.}
}
\value{
\code{bcai()} takes two vectors and returns Benati and Capurri's (2026)
alignment index (A).
}
\description{
\code{bcai()} takes two vectors and returns Benati and Capurri's (2026)
alignment index (A).
}
\details{
You can think of the alignment index that Benati and Capurri (2026) describe
as an \emph{S} corollary to the chance-corrected measures that Häge (2011) offers
as substitutes for \emph{S}. It takes the (unweighted, absolute distances) \emph{S}
score proposed by Signorino and Ritter (1999) and subtracts from it the \emph{S}
score that would follow under the assumption of independent voting.

The function subsets to complete cases of the two vectors for which you want
an alignment score.

The function implicitly assumes that \code{x1} and \code{x2} are columns in a data
frame. One indirect check for this looks at whether \code{x1} and \code{x2} are the
same length. The function will stop if they're not.

There will sometimes be instances, assuredly with alliances, where not all
categories are observed. For example, the toy example I provide of Germany
and Russia in 1914 includes no 2s. In the language of "ratings", the "rating"
of 2 was available for Germany and Russia in 1914 but neither side used it.
The \code{levels} argument allows you to specify the full sequence of values that
could be observed, even if none were. It probably makes the most sense to
always use this argument, even if the default behavior operates as if you
won't.
\subsection{A Few Caveats on Weighting}{

You can weight this measure if you want. Please be mindful about what you're
doing, especially if the weights are CINC scores. See here:

\url{https://svmiller.com/blog/2026/06/alliances-weighting-foreign-policy-similarity/}

The function will proportionalize your weights to sum to 1 if they do not
already.
}
}
\examples{

bcai(gmyrus14$gmy, gmyrus14$rus, levels = 0:3) # with levels argument
bcai(bencapex$rowv, bencapex$colv) # levels argument not necessary here.
bcai(gmyrus14$gmy, gmyrus14$rus, distances = 'squared', levels = 0:3) # squared, with levels argument

}
\references{
Benati, Stefano, and Agnese Capurri. 2026. "The Alignment index and its
application to voting at the United Nations General Assembly." \emph{Quality &
Quantity}. \doi{10.1007/s11135-026-02814-x}
}
